Used typescript in the front-end Relates #73
