+++ client/src/App/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './index.css';
+
+function App() {
+  return (
+    <div>
+      Hello, Quizzer!
+    </div>
+  );
+}
+
+export default App;
