+++ client/src/mui.ts
@@ -1,3 +1,4 @@
+import { StyledEngineProvider } from '@mui/material';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 
@@ -5,4 +6,5 @@ export {
   CssBaseline,
   ThemeProvider,
   createTheme,
+  StyledEngineProvider,
 };
