+++ client/src/typings/declaration.d.ts
@@ -1,0 +1,1 @@
+declare module '*.module.css';
